+++ src/deployment/RuleEngineOwnable.sol
@@ -25,6 +25,7 @@ contract RuleEngineOwnable is RuleEngineOwnableShared, Ownable {
      * @dev Access control check using Ownable pattern
      */
     function _onlyRulesManager() internal virtual override onlyOwner {}
+    function _onlyRulesLimitManager() internal virtual override onlyOwner {}
 
     /**
      * @dev Access control check using Ownable pattern