+++ test/RuleEngine/AccessControl/RuleEngineAccessControl.sol
@@ -95,4 +95,23 @@ contract RuleEngineTest is Test, HelperContract {
         vm.expectRevert(ERC3643ComplianceModule.RuleEngine_ERC3643Compliance_UnauthorizedCaller.selector);
         ruleEngineMock.transferred(address(0), ADDRESS1, ADDRESS2, 10);
     }
+
+    function testCannotAttackerSetMaxRules() public {
+        vm.prank(ATTACKER);
+        vm.expectRevert(
+            abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, ATTACKER, bytes32(0))
+        );
+        ruleEngineMock.setMaxRules(12);
+    }
+
+    function testRulesManagerCannotSetMaxRulesWithoutAdminRole() public {
+        vm.prank(RULE_ENGINE_OPERATOR_ADDRESS);
+        ruleEngineMock.grantRole(RULES_MANAGEMENT_ROLE, WHITELIST_OPERATOR_ADDRESS);
+
+        vm.prank(WHITELIST_OPERATOR_ADDRESS);
+        vm.expectRevert(
+            abi.encodeWithSelector(AccessControlUnauthorizedAccount.selector, WHITELIST_OPERATOR_ADDRESS, bytes32(0))
+        );
+        ruleEngineMock.setMaxRules(12);
+    }
 }